requirements.txt
streamlit
python-dotenv

langchain
langchain-core
langchain-community
langchain-text-splitters

langchain-google-genai

chromadb
fastembed

pypdf